"Shifted Peaks" Simulated Dataset

A simulated dataset generated for examples in the veesa pipeline manuscript. For the code used to prepare this dataset, see https://github.com/sandialabs/veesa/inst/data-shifted-peaks.md.

Usage

shifted_peaks

Format

A list.

Details

The objects in the list are:

data	Data frame containing simulated data

params	The parameters used to generate the data

true_means	The true functional means of the shifted peaks groups.